+++ src/proto/streams/mod.rs
@@ -33,9 +33,6 @@ use std::time::Duration;
 
 #[derive(Debug)]
 pub struct Config {
-    /// Initial window size of locally initiated streams
-    pub local_init_window_sz: WindowSize,
-
     /// Initial maximum number of locally initiated streams.
     /// After receiving a Settings frame from the remote peer,
     /// the connection will overwrite this value with the
@@ -69,4 +66,10 @@ pub struct Config {
 
     /// Maximum number of remote initiated streams
     pub remote_max_initiated: Option<usize>,
+
+    /// Maximum number of locally reset streams due to protocol error across
+    /// the lifetime of the connection.
+    ///
+    /// When this gets exceeded, we issue GOAWAYs.
+    pub local_max_error_reset_streams: Option<usize>,
 }